Refactor: rename + extract widget in separate files
