+++ lib/ui/viewer/gallery/component/grid/recyclable_grid_view_widget.dart
@@ -1,12 +1,12 @@
 import "package:flutter/material.dart";
 import "package:photos/models/file.dart";
 import "package:photos/models/selected_files.dart";
-import "package:photos/ui/huge_listview/place_holder_widget.dart";
-import "package:photos/ui/viewer/gallery/component/lazy_loading_gallery.dart";
+import "package:photos/ui/viewer/gallery/component/grid/gallery_grid_vew_widget.dart";
+import "package:photos/ui/viewer/gallery/component/grid/place_holder_grid_view_widget.dart";
 import "package:photos/ui/viewer/gallery/gallery.dart";
 import "package:visibility_detector/visibility_detector.dart";
 
-class RecyclableViewWidget extends StatefulWidget {
+class RecyclableGridViewWidget extends StatefulWidget {
   final bool shouldRender;
   final List<File> filesInDay;
   final int photoGridSize;
@@ -15,7 +15,7 @@ class RecyclableViewWidget extends StatefulWidget {
   final GalleryLoader asyncLoader;
   final int? currentUserID;
   final SelectedFiles? selectedFiles;
-  const RecyclableViewWidget({
+  const RecyclableGridViewWidget({
     required this.shouldRender,
     required this.filesInDay,
     required this.photoGridSize,
@@ -28,10 +28,11 @@ class RecyclableViewWidget extends StatefulWidget {
   });
 
   @override
-  State<RecyclableViewWidget> createState() => _RecyclableViewWidgetState();
+  State<RecyclableGridViewWidget> createState() =>
+      _RecyclableGridViewWidgetState();
 }
 
-class _RecyclableViewWidgetState extends State<RecyclableViewWidget> {
+class _RecyclableGridViewWidgetState extends State<RecyclableGridViewWidget> {
   late bool _shouldRender;
   @override
   void initState() {
@@ -61,7 +62,10 @@ class _RecyclableViewWidgetState extends State<RecyclableViewWidget> {
               selectedFiles: widget.selectedFiles,
               currentUserID: widget.currentUserID,
             )
-          : PlaceHolderWidget(widget.filesInDay.length, widget.photoGridSize),
+          : PlaceHolderGridViewWidget(
+              widget.filesInDay.length,
+              widget.photoGridSize,
+            ),
     );
   }
 }
